--- src/student_aleksandrs_jakovenko/lesson_3/level_1/Robot.java
@@ -1,14 +1,0 @@
-package student_aleksandrs_jakovenko.lesson_3.level_1;
-
-public class Robot {
-    public void sayHello(){
-        System.out.println("Hello!");
-    }
-
-    public void sayYourName() {
-        System.out.println("My name is ROBO");
-    }
-
-    public Robot() {
-    }
-}
